requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-peewee==3.0.3
Flask-WTF==0.14.3
geographiclib==1.50
geoip2==3.0.0
geopy==1.22.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
maxminddb==1.5.4
peewee==3.13.1
pyactiveresource==2.1.2
PyMySQL==0.9.3
PyYAML==5.3
requests==2.24.0
ShopifyAPI==5.1.2
six==1.13.0
urllib3==1.25.9
Werkzeug==0.16.0
wtf-peewee==3.0.0
WTForms==2.2.1
